Move `setNumericValues` to `actions/utils`
